dagster/requirements.txt
dagster
dagster-webserver
boto3==1.37.38
dbt-core==1.9.0
dbt-postgres==1.9.0
opentelemetry-api==1.28.2
opentelemetry-exporter-otlp-proto-http==1.28.2
opentelemetry-sdk==1.28.2
prometheus-client==0.21.1
psycopg[binary]==3.2.5
requests==2.32.3